+++ modules/shared_items.py
@@ -23,3 +23,18 @@ def list_crossattention():
         "Sub-quadratic",
         "Split attention"
     ]
+
+def get_pipelines():
+    import diffusers
+    return {
+        'Autodetect': None,
+        'Stable Diffusion': diffusers.StableDiffusionPipeline,
+        'Stable Diffusion Img2Img': diffusers.StableDiffusionImg2ImgPipeline,
+        'Stable Diffusion Instruct': diffusers.StableDiffusionInstructPix2PixPipeline,
+        'Stable Diffusion Upscale': diffusers.StableDiffusionUpscalePipeline,
+        'Stable Diffusion XL': diffusers.StableDiffusionXLPipeline,
+        'Stable Diffusion XL Img2Img': diffusers.StableDiffusionXLImg2ImgPipeline,
+        'Stable Diffusion XL Inpaint': diffusers.StableDiffusionXLInpaintPipeline,
+        'Stable Diffusion XL Instruct': diffusers.StableDiffusionXLInstructPix2PixPipeline,
+        # 'Kandinsky V1', 'Kandinsky V2', 'DeepFloyd IF', 'Shap-E', 'Kandinsky V1 Img2Img', 'Kandinsky V2 Img2Img', 'DeepFloyd IF Img2Img', 'Shap-E Img2Img',
+    }